test user info for booking tickets

+++ 619Project/src/CustomerModel/CustomerManager.java
@@ -47,7 +47,7 @@ public class CustomerManager {
     }
     
     // for registration, call processTransaction() first before calling registerUser() 
-    public RegisteredUser registerUser(String name, String addr, String bank, int card_no, int expr_date, int cvv, String email, String username, String pwd) {
+    public RegisteredUser registerUser(String name, String addr, String bank, long card_no, int expr_date, int cvv, String email, String username, String pwd) {
     	RegisteredUser newUser = new RegisteredUser(name, addr, bank, card_no, expr_date, cvv, email, username, pwd);
     	newUser.paidAnnualFee();
     	registeredUsers.put(username, newUser);
